--- test/nodes/security/requirements.txt
@@ -1,3 +1,0 @@
-hypothesis>=6.0
-jsonschema>=4.0
-pytest>=7.0
